refactor: move data layer interfaces to folders
