+++ src/data/usecases/add-account/db-add-account-protocols.ts
@@ -2,5 +2,5 @@
 
 export * from '../../../domain/models/account'
 export * from '../../../domain/usecases/add-account'
-export * from '../../protocols/encrypter'
-export * from '../../protocols/add-account-repository'
+export * from '../../protocols/criptography/encrypter'
+export * from '../../protocols/db/add-account-repository'
