+++ src/data/usecases/authentication/db-authentication.ts
@@ -2,7 +2,7 @@ import {
   Authentication,
   AuthenticationModel,
 } from '../../../domain/usecases/authentication'
-import { LoadAccountByEmailRepository } from '../../protocols/load-account-by-email-repository'
+import { LoadAccountByEmailRepository } from '../../protocols/db/load-account-by-email-repository'
 
 export class DbAuthentication implements Authentication {
   constructor(
